@@ -7,14 +7,14 @@ plugins {
   kotlin("android")
 }
 
-val javaVersion = JavaVersion.VERSION_1_8
+val javaVersion = JavaVersion.VERSION_11
 
 android {
+  namespace = "de.mannodermaus.junit5.testutil"
   compileSdk = Android.compileSdkVersion
 
   defaultConfig {
     minSdk = 4
-    targetSdk = Android.targetSdkVersion
   }
 
   compileOptions {
@@ -29,16 +29,18 @@ android {
 
   lint {
     // JUnit 4 refers to java.lang.management APIs, which are absent on Android.
-    warning("InvalidPackage")
+    warning.add("InvalidPackage")
+    targetSdk = Android.targetSdkVersion
   }
 
-  packagingOptions {
+  packaging {
     resources.excludes.add("META-INF/LICENSE.md")
     resources.excludes.add("META-INF/LICENSE-notice.md")
   }
 
   testOptions {
     unitTests.isReturnDefaultValues = true
+    targetSdk = Android.targetSdkVersion
   }
 }
 
