@@ -7,17 +7,17 @@ import androidx.compose.foundation.layout.Column
 import androidx.compose.material.Button
 import androidx.compose.material.Text
 import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.mutableIntStateOf
 import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
 
-class ExistingActivity : ComponentActivity() {
+public class ExistingActivity : ComponentActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
 
         setContent {
             Column {
-                var counter by remember { mutableStateOf(0) }
+                var counter by remember { mutableIntStateOf(0) }
 
                 Text(text = "Clicked: $counter")
                 Button(onClick = { counter++ }) {
